Move prometheusMiddleware to middleware package
